Implementation core entities for lesson service.
The lesson consists of the following attributes: title, description, text content of the lesson (theoretical part), links to the image for the lesson card, task list.
The code snippet is one of the optional attributes of the lesson. Consists of an explanation of the code, the code itself, and the type of language in which the code is written.